feat(core): add proseMathMarkdownSyntaxExtension to default markdown
Move dollar-delimited math parsing to core as Math/MathMark/MathFormula nodes,
include in prosemarkMarkdownSyntaxExtensions, add @lezer/highlight dependency.

Latex package re-exports latex* aliases, folds Math nodes with hybrid block rules.
Demos: drop redundant latex syntax import. Docs: mention prose math + latex widgets.
