+++ apps/docs/src/components/ProseMarkDemo.astro
@@ -23,7 +23,6 @@
   import {
     latexMarkdownEditorExtensions,
     latexMarkdownSyntaxTheme,
-    latexMathMarkdownSyntaxExtension,
   } from '@prosemark/latex';
   import {
     pasteRichTextExtension,
@@ -44,7 +43,6 @@
           GFM,
           prosemarkMarkdownSyntaxExtensions,
           renderHtmlMarkdownSyntaxExtensions,
-          latexMathMarkdownSyntaxExtension,
         ],
       }),
       prosemarkBasicSetup(),
