+++ packages/core/package.json
@@ -32,13 +32,13 @@
   },
   "dependencies": {
     "@lezer/common": "^1.4.0",
+    "@lezer/highlight": "^1.2.3",
     "@lezer/markdown": "^1.6.1",
     "@lezer/yaml": "^1.0.4",
     "ajv": "^8.17.1",
     "node-emoji": "^2.2.0"
   },
   "peerDependencies": {
-    "@lezer/highlight": "^1.2.0",
     "@codemirror/autocomplete": "^6.20.0",
     "@codemirror/commands": "^6.10.1",
     "@codemirror/lang-markdown": "^6.5.0",
